requirements.txt
# Core dependencies for site_crawler.py and generate_markdown.py

# Web crawling
crawl4ai>=0.3.0

# Database ORM
sqlmodel>=0.0.14

# Environment variables
python-dotenv>=1.0.0

# LangGraph and OpenAI dependencies for list_all_cpe_products.py
langgraph>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
pydantic>=1.10.0

